Avoid waking some constraints too often
When a lower bound changes from k to k+p, all literals
from ~(x<=k-1) to ~(x<=k+p-1) are placed in the trail.
This change makes it so we only wake the propagator
when the literal we are processing matches the actual
bound of the variable, but we always invoke it once
when a bound changes.

Note this might delay waking a constraint: if a propagator
sets x >= 5, then another sets x >= 10 before we get to
processing x>=5, all propagators that are watching the
lower bound of x will wake only when we process x >= 10,
not for both literals.