Throwing was the wrong approach in cspvar::*f()
Instead of throwing, we simply construct a clause without
filling in the missing literal and return that as a
conflict.

The reasoning is simple: there are implicitly an
infinite number of <= variables for each domain,
all of which below x.min() are false and all
above x.max() are true at level 0. So the clause
minus that (implicit) literal is still a valid
clause. Same of course for =.