Free inactive clauses.
There a truly humongous number of inactive clauses generated,
eating all memory pretty quickly. It is not clear to me how often
we need to garbage collect, maybe it will be easier when I do
the port to minisat 2.2.

By the way, this provides a measurable speedup. Profiles
point to malloc, at least partially. This is probably because
most clauses generated by constraints are small so malloc
gets to reuse space from previous allocations.