Increase unit growth limit for inlining. 5-10% speedup.
It was annoying that vec<Lit>::push kept showing up on profiles.
This makes it succesfully inline which improves performance by ~5-10%,
but the real cause is eagerly generating clauses during pruning. Much
of the cost of cons_neq, for example, is generating the clause and
the associated malloc.

When this is fixed, I expect a more significant speedup.