Dynamically split alldiff at runtime.
This is based on the Gent et al paper, but
unlike their results, it doesn't seem to help
here. If anything, it's marginally slower. Still,
I leave it because I might not have enough
benchmarks yet or there might be a bug that
I will discover later. In any case, it is
easy enough to revert back to the original
code if needed.